helpers: Rename file http_gate.py (HTTP -> REST)
