+++ pytest_tests/lib/http_gw/http_utils.py
@@ -2,7 +2,7 @@ import random
 
 from neofs_testlib.env.env import StorageNode
 from neofs_testlib.shell import Shell
-from helpers.http_gate import assert_hashes_are_equal, get_via_rest_gate
+from helpers.rest_gate import assert_hashes_are_equal, get_via_rest_gate
 from helpers.neofs_verbs import get_object
 from helpers.complex_object_actions import get_nodes_without_object
