+++ pytest_tests/tests/services/rest_gate/test_rest_streaming.py
@@ -4,7 +4,7 @@ import allure
 import pytest
 from helpers.container import create_container
 from helpers.file_helper import generate_file
-from helpers.http_gate import upload_via_rest_gate
+from helpers.rest_gate import upload_via_rest_gate
 from helpers.wellknown_acl import PUBLIC_ACL
 from http_gw.http_utils import get_object_and_verify_hashes
 from neofs_env.neofs_env_test_base import NeofsEnvTestBase
